fixed a few mobile style issues.
tuned down the grid collumn minwidth to not overflow screens 360px wide.
made up for the diff in collumn gap to keep 4 per row on FHD-landscape.

added periods to soc media icon link text for that weird instance when
there's no style and these links are 'sentences' in a <p>